refactor(control-plane): rehome internal capability kernels
