--- loopx/capabilities/review_batch/__init__.py
@@ -1,11 +1,0 @@
-"""Provider-neutral review-batch composition and decision binding."""
-
-from .core import (
-    build_review_batch,
-    bind_review_batch_decisions,
-)
-
-__all__ = [
-    "bind_review_batch_decisions",
-    "build_review_batch",
-]
