+++ loopx/cli_commands/__init__.py
@@ -104,6 +104,7 @@ from .presentation import handle_presentation_command, register_presentation_com
 from .pr_review import handle_pr_review_command, register_pr_review_command
 from .quota import handle_quota_command, register_quota_command
 from .ready_score import handle_ready_score_command, register_ready_score_command
+from .review_batch import handle_review_batch_command, register_review_batch_commands
 from .registry_admin import (
     handle_registry_admin_command,
     register_registry_admin_commands,
@@ -178,6 +179,7 @@ from .version import handle_version_command, register_version_command
 from .worker_bridge import handle_worker_bridge_command, register_worker_bridge_commands
 
 __all__ = [
+    "build_lark_issue_fix_reviewer_provider_hooks",
     "handle_agents_last_exam_command",
     "handle_agents_last_exam_baked_input_command",
     "handle_agents_last_exam_host_codex_command",
@@ -235,6 +237,7 @@ __all__ = [
     "handle_pr_review_command",
     "handle_quota_command",
     "handle_ready_score_command",
+    "handle_review_batch_command",
     "handle_registry_admin_command",
     "handle_review_packet_command",
     "handle_slash_commands_command",
@@ -289,6 +292,7 @@ __all__ = [
     "register_presentation_commands",
     "register_quota_command",
     "register_ready_score_command",
+    "register_review_batch_commands",
     "register_registry_admin_commands",
     "register_slash_commands_command",
     "register_starter_commands",
